app/requirements.txt
poetry==1.7.0
jwt==1.3.1
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose==3.3.0
dotenv
